+++ src/core/error.rs
@@ -21,6 +21,8 @@ pub const CODE_PERMISSION_DENIED: i32 = 13;
 pub enum SourceError {
     #[error(transparent)]
     Io(#[from] std::io::Error),
+    #[error(transparent)]
+    Ureq(#[from] ureq::Error),
     #[cfg(feature = "serde")]
     #[error(transparent)]
     SerdeJson(#[from] ::serde_json::Error),