+++ src/lib.rs
@@ -2,6 +2,7 @@ pub use core::error::*;
 pub use core::host::Host;
 pub use core::log::*;
 pub use core::metrics::Metrics;
+pub use core::target::{Scheme, Target};
 
 pub mod connectors;
 pub mod core;